Feature: display book cards
Added some HTML to have some sections and a div that contains the book
cards. The css it's light for the moment.

For javascript, created an array to keep the Book objects. A Book object
has title, author, page count and read properties. There is a function
that creates some book cards when loading the page.

Another functions are to create the elements needed for a book card,
to add a book to the array and to display each book card.