Feature: add new books
There is a new button in the page that opens a form when clicked. It
lets the user enter data for a new book to display. The form has two
buttons, one to hide and reset the form fields and the other to submit
the form. This form shows at the center of the page.

For the moment, there is no validation on the form fields. Javascript
gets the form data, creates a Book object and adds it to both the
myLibrary array and the page.

There is some styling to the form, that is inside a div. Also the footer
now stays at the bottom of the page.