Feature: add delete buttons
Now each book card has a button to delete it. It also deletes the Book
object in myLibrary array.

To do this, each book card has a data attribute that's the index of the
book in myLibrary. It uses a method to remove it from the array and
call another method on the book card to remove it. After this, it calls
a function to update the indexes of the remaining book cards.