Feature: add toggle button to book cards
There is now a button to toggle the read status of a Book object. It's
created using javascript. It consists of a label and a checkbox and span
elements inside the label. The css code gives the look of a typical
toggle button.

When clicking the new book button, the delete and toggle buttons get
disabled until the user creates or clicks the cancel button in the form.

Commented the prefillLibrary and displayLibrary functions. This was
generating a bug where the new book cards weren't sincronized with the
premade book cards. Also, changed the functions that create the buttons
to now show them as disabled.

The book cards now use flex. The buttons are now bigger.

Fixed a typo in the grid definition.