Refactor: use class syntax for Book and myLibrary
Created new Book class. It has a constructor function and the
toggleRead function.

Created new Library class. Added books array and getter method. Also
added addBookToLibrary method.

Added a Placeholder class with the intention of having a class that
handles the logic of creating, deleting and updating DOM elements. For
the moment it has some variables that reference some of the elements
used in many global functions.

Fixed indentation in one function. Added let to some variables in for
loops.

In HTML, changed the ids to classes. The input elements from the form
are now required to prevent empty book cards.

In CSS, modified some rules to keep applying styles correclty.